requirements.txt
python-dotenv>=1.0.1
playwright>=1.52.0
psycopg[binary]>=3.2.9
requests>=2.32.3
sentence-transformers>=3.0.1
fastapi>=0.116.1
uvicorn>=0.35.0
